🔵 move some testing functions to `commandsFileGeneration.go`
